requirements.txt
openai
pygithub
python-dotenv
termcolor
